Renamed arm_ros_firmware to umrt-arm-ros-firmware. Note that C++ namespaces can't have hyphens, so have been left as umrt_arm_ros_firmware.
