--- arm-ros-firmware/.idea/.name
@@ -1,1 +1,0 @@
-arm_firmware
