+++ umrt-arm-ros-firmware/description/urdf/diffbot.urdf.xacro
@@ -4,13 +4,13 @@
   <xacro:arg name="prefix" default="" />
   <xacro:arg name="use_mock_hardware" default="false" />
 
-  <xacro:include filename="$(find arm_ros_firmware)/ros2_control_demo_description/diffbot/urdf/diffbot_description.urdf.xacro" />
+  <xacro:include filename="$(find umrt-arm-ros-firmware)/ros2_control_demo_description/diffbot/urdf/diffbot_description.urdf.xacro" />
 
   <!-- Import Rviz colors -->
-  <xacro:include filename="$(find arm_ros_firmware)/ros2_control_demo_description/diffbot/urdf/diffbot.materials.xacro" />
+  <xacro:include filename="$(find umrt-arm-ros-firmware)/ros2_control_demo_description/diffbot/urdf/diffbot.materials.xacro" />
 
   <!-- Import diffbot ros2_control description -->
-  <xacro:include filename="$(find arm_ros_firmware)/ros2_control/diffbot.ros2_control.xacro" />
+  <xacro:include filename="$(find umrt-arm-ros-firmware)/ros2_control/diffbot.ros2_control.xacro" />
 
   <xacro:diffbot prefix="$(arg prefix)" />
