+++ umrt-arm-ros-firmware/hardware/include/umrt-arm-ros-firmware/visibility_control.h
@@ -19,38 +19,38 @@
  * library cannot have, but the consuming code must have inorder to link.
  */
 
-#ifndef ARM_ROS_FIRMWARE__VISIBILITY_CONTROL_H_
-#define ARM_ROS_FIRMWARE__VISIBILITY_CONTROL_H_
+#ifndef UMRT_ARM_ROS_FIRMWARE__VISIBILITY_CONTROL_H_
+#define UMRT_ARM_ROS_FIRMWARE__VISIBILITY_CONTROL_H_
 
 // This logic was borrowed (then namespaced) from the examples on the gcc wiki:
 //     https://gcc.gnu.org/wiki/Visibility
 
 #if defined _WIN32 || defined __CYGWIN__
 #ifdef __GNUC__
-#define ARM_ROS_FIRMWARE_EXPORT __attribute__((dllexport))
-#define ARM_ROS_FIRMWARE_IMPORT __attribute__((dllimport))
+#define UMRT_ARM_ROS_FIRMWARE_EXPORT __attribute__((dllexport))
+#define UMRT_ARM_ROS_FIRMWARE_IMPORT __attribute__((dllimport))
 #else
-#define ARM_ROS_FIRMWARE_EXPORT __declspec(dllexport)
-#define ARM_ROS_FIRMWARE_IMPORT __declspec(dllimport)
+#define UMRT_ARM_ROS_FIRMWARE_EXPORT __declspec(dllexport)
+#define UMRT_ARM_ROS_FIRMWARE_IMPORT __declspec(dllimport)
 #endif
-#ifdef ARM_ROS_FIRMWARE_BUILDING_DLL
-#define ARM_ROS_FIRMWARE_PUBLIC ARM_ROS_FIRMWARE_EXPORT
+#ifdef UMRT_ARM_ROS_FIRMWARE_BUILDING_DLL
+#define UMRT_ARM_ROS_FIRMWARE_PUBLIC UMRT_ARM_ROS_FIRMWARE_EXPORT
 #else
-#define ARM_ROS_FIRMWARE_PUBLIC ARM_ROS_FIRMWARE_IMPORT
+#define UMRT_ARM_ROS_FIRMWARE_PUBLIC UMRT_ARM_ROS_FIRMWARE_IMPORT
 #endif
-#define ARM_ROS_FIRMWARE_PUBLIC_TYPE ARM_ROS_FIRMWARE_PUBLIC
-#define ARM_ROS_FIRMWARE_LOCAL
+#define UMRT_ARM_ROS_FIRMWARE_PUBLIC_TYPE UMRT_ARM_ROS_FIRMWARE_PUBLIC
+#define UMRT_ARM_ROS_FIRMWARE_LOCAL
 #else
-#define ARM_ROS_FIRMWARE_EXPORT __attribute__((visibility("default")))
-#define ARM_ROS_FIRMWARE_IMPORT
+#define UMRT_ARM_ROS_FIRMWARE_EXPORT __attribute__((visibility("default")))
+#define UMRT_ARM_ROS_FIRMWARE_IMPORT
 #if __GNUC__ >= 4
-#define ARM_ROS_FIRMWARE_PUBLIC __attribute__((visibility("default")))
-#define ARM_ROS_FIRMWARE_LOCAL __attribute__((visibility("hidden")))
+#define UMRT_ARM_ROS_FIRMWARE_PUBLIC __attribute__((visibility("default")))
+#define UMRT_ARM_ROS_FIRMWARE_LOCAL __attribute__((visibility("hidden")))
 #else
-#define ARM_ROS_FIRMWARE_PUBLIC
-#define ARM_ROS_FIRMWARE_LOCAL
+#define UMRT_ARM_ROS_FIRMWARE_PUBLIC
+#define UMRT_ARM_ROS_FIRMWARE_LOCAL
 #endif
-#define ARM_ROS_FIRMWARE_PUBLIC_TYPE
+#define UMRT_ARM_ROS_FIRMWARE_PUBLIC_TYPE
 #endif
 
-#endif  // ARM_ROS_FIRMWARE__VISIBILITY_CONTROL_H_
+#endif  // UMRT_ARM_ROS_FIRMWARE__VISIBILITY_CONTROL_H_
