+++ umrt-arm-ros-firmware/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>arm_ros_firmware</name>
+  <name>umrt-arm-ros-firmware</name>
   <version>0.3.0</version>
   <description>Demo package of `ros2_control` hardware for DiffBot.</description>
