+++ umrt-arm-ros-firmware/test/test_urdf_xacro.py
@@ -38,7 +38,7 @@ from ament_index_python.packages import get_package_share_directory
 
 def test_urdf_xacro():
     # General Arguments
-    description_package = "arm_ros_firmware"
+    description_package = "umrt-arm-ros-firmware"
     description_file = "diffbot.urdf.xacro"
 
     description_file_path = os.path.join(
