+++ umrt-arm-ros-firmware/umrt-arm-ros-firmware.xml
@@ -1,6 +1,6 @@
-<library path="arm_ros_firmware">
-  <class name="arm_ros_firmware/DiffBotSystemHardware"
-         type="arm_ros_firmware::DiffBotSystemHardware"
+<library path="umrt-arm-ros-firmware">
+  <class name="umrt-arm-ros-firmware/DiffBotSystemHardware"
+         type="umrt_arm_ros_firmware::DiffBotSystemHardware"
          base_class_type="hardware_interface::SystemInterface">
     <description>
       The ros2_control DiffBot example using a system hardware interface-type. It uses velocity command and position state interface. The example is the starting point to implement a hardware interface for differential-drive mobile robots.
